pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "reviewer-two-env"
version = "0.1.0"
description = "It's Reviewer Two, but it's an OpenEnv."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Chris von Csefalvay", email = "chris@chrisvoncsefalvay.com"},
]
keywords = [
    "reinforcement-learning",
    "openenv",
    "research",
    "llm",
    "evaluation",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "openenv-core @ git+https://github.com/meta-pytorch/OpenEnv.git",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "datasets>=2.14.0",
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "sentencepiece>=0.1.99",
    "accelerate>=0.25.0",
    "nltk>=3.8.0",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
green-agent = [
    "a2a-sdk[http-server]>=0.3.20",
    "httpx>=0.28.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.28.0",
]

[project.scripts]
research-plan-green-agent = "research_plan_env.green_agent.server:main"

[project.urls]
Homepage = "https://github.com/chrisvoncsefalvay/reviewer-two-env"
Repository = "https://github.com/chrisvoncsefalvay/reviewer-two-env"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["research_plan_env"]
